Fixed Map drawing issue in Apia Central [#156 responsible:none state:fixed]
Issue was due to a parcel being reset with an empty firstpart and lastpart
using the new Change Parcel Atributes tool. The map was unable to calculate
the position of parcel label (i.e. the \n Newline) for display the map.
Fixed by changing the cadastre.formatparcelnrlabel proc to omit the \n
if either firstpart or lastpart are empty. Also updated the change_parcel_name
stored proc to prevent the issue from re-occuring.